# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel 12 API Documentation",
        "description": "API Documentation for Laravel 12 Application",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/weather/perth": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Fetch weather data for Perth, Australia",
                "description": "Retrieves the current weather data for Perth using OpenWeatherMap API.",
                "operationId": "a57f5de992a68e33c1ba26bb09565d07",
                "responses": {
                    "200": {
                        "description": "Weather data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "temperature": {
                                                    "type": "number",
                                                    "example": 22.5
                                                },
                                                "feels_like": {
                                                    "type": "number",
                                                    "example": 21.8
                                                },
                                                "humidity": {
                                                    "type": "integer",
                                                    "example": 60
                                                },
                                                "weather": {
                                                    "type": "string",
                                                    "example": "clear sky"
                                                },
                                                "wind_speed": {
                                                    "type": "number",
                                                    "example": 5.1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch weather data"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "description": "Use a bearer token to access this API",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "BearerToken": []
        }
    ]
}